some stuff

+++ lucid/lucid/math/animation.cpp
@@ -1,15 +1,15 @@
 #include "../lucid.h"
 
-double func_linear(double x, double to, double speed_multiplier) {
+double func_linear(float x, float to, float speed_multiplier) {
 	return std::clamp(x * lucid_engine::g_io->m_delta_time * speed_multiplier, x, to);
 }
 
-double func_lerp(double from, double to, double speed_multiplier) {
+double func_lerp(float from, float to, float speed_multiplier) {
 	return from + (to - from) * lucid_engine::g_io->m_delta_time * speed_multiplier;
 }
 
-double lucid_engine::math::animate(animation_type animation, double from, double to, double speed_multiplier) {
-	double x = 0.0;
+float lucid_engine::math::animate(animation_type animation, float from, float to, float speed_multiplier) {
+	float x = 0.0;
 
 	switch (animation) {
 	case animation_type::linear:
